move allocation of Environment into safe wrapper
